move hook to hook folder ahoy i lost me hand but now i have a hook
